requirements.txt
https://github.com/Rapptz/discord.py/archive/async.zip
PyYAML==3.11